refactor(firestore-send-email): additional updates
