+++ firestore-send-email/functions/jest.config.js
@@ -1,4 +1,4 @@
-const packageJson = require("./functions/package.json");
+const packageJson = require("./package.json");
 
 module.exports = {
   name: packageJson.name,
