+++ firestore-send-email/functions/jest.teardown.js
