+++ firestore-send-email/functions/package.json
@@ -1,13 +1,13 @@
 {
-  "name": "firestore-send-email",
+  "name": "firestore-send-email-functions",
   "description": "",
   "main": "lib/index.js",
   "scripts": {
     "build": "npm run clean && npm run compile",
     "clean": "rimraf lib",
     "compile": "tsc",
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "generate-readme": "firebase ext:info . --markdown > README.md"
+    "test": "jest",
+    "generate-readme": "firebase ext:info .. --markdown > README.md"
   },
   "keywords": [],
   "author": "",
