--- firestore-send-email/tsconfig.json
@@ -1,7 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "outDir": "functions/lib"
-  },
-  "include": ["functions/src"]
-}
