+++ lerna.json
@@ -1,4 +1,8 @@
 {
-  "packages": ["*", "firestore-translate-text/functions"],
+  "packages": [
+    "*",
+    "firestore-translate-text/functions",
+    "firestore-send-email/functions"
+  ],
   "version": "0.0.0"
 }
